requirements.txt
bing_image_urls~=0.1.5
discord.py==2.3.2
python-dotenv==1.0.0
requests~=2.28.2
twilio==7.16.4
yfinance==0.2.12
openai~=0.27.0